removed hard coded VSGDB from Venus !!!! minor fixes to volutil (more coming) numbers for mond.rpc2

+++ coda-src/vice/srvproc.cc
@@ -4822,7 +4822,8 @@ static void Perform_RR(ClientEntry *client, VolumeId VSGVolnum, Volume *volptr,
 
     /* Await COP2 message. */
     if (ReplicatedOp) {
-	    ViceFid fids[MAXFIDS]; bzero((void *)fids, (int) (MAXFIDS * sizeof(ViceFid)));
+	    ViceFid fids[MAXFIDS]; 
+	    bzero((void *)fids,  (MAXFIDS * sizeof(ViceFid)));
 	    fids[0] = Did;
 	    SLog(3, "Perform_RR: delete_me = %d, !delete_me = %d",
 		 vptr->delete_me, !vptr->delete_me);
